% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votempls.R
\docType{package}
\name{votempls}
\alias{votempls}
\alias{votempls-package}
\title{votempls: Minneapolis ranked choice voting results.}
\description{
This dataset contains the ranked choice ballots for the 2013 and 2017
Minneapolis elections. It includes the elections for mayor, city council,
park and recreation commissioner, park board, and board of estimate and taxation.
}
